Add a default toString method
So this adds a very benifical toString method, helpful when inspecting
the tagged and taggedSum. There is an issue and that issue is the fact
that we don't have the name of the tagged item, but I'm not sure if that
even matters sometimes.

Optionally you can also override the default toString and implement your
own one.